requirements.txt
PyQt5==5.15.9
Pillow>=9.5.0
PyQt5-sip<13.0.0,>=12.11.0
PyQt5-Qt5>=5.15.2
